create activity.mails.js and moved mails from methods and update imports
